+++ db/element/edge.h
@@ -28,7 +28,7 @@ namespace element
     class edge : public element
     {
         public:
-            edge(uint64_t time, int remote_loc, uint64_t remote_handle);
+            edge(uint64_t time, uint64_t remote_loc, uint64_t remote_handle);
             edge(uint64_t time, remote_node rn);
         
         public:
@@ -36,7 +36,7 @@ namespace element
     };
 
     inline
-    edge :: edge(uint64_t time, int remote_loc, uint64_t remote_handle)
+    edge :: edge(uint64_t time, uint64_t remote_loc, uint64_t remote_handle)
         : element(time)
         , nbr(remote_loc, remote_handle)
     {
